chore: remove redundant .env files
Remove deployment artifact .env files that are created on-demand:
- .env.email: generated by deploy_email.sh (not required in git)
- .env.mailu: generated by deploy_simple_mail.py (not required in git)
- .env.simple-mail: generated by deploy_email.sh (not required in git)

Retains:
- .env (active configuration with real secrets, not committed)
- .env.consolidated (template with placeholders, tracked in git)

Also fix: Updated pre-commit hook to allow deletion of .env files
(only blocks ADD/MODIFy of .env files with secrets)

+++ linkedin_job_scanner.py
@@ -96,7 +96,7 @@ JOB_APPLY_TIMEOUT = 90  # Timeout máximo (segundos) por candidatura individual
 
 # ─── Pretensão Salarial Dinâmica (CLT mensal BRL) ────────────────────────────
 # Faixas base por nível de senioridade (min, max)
-SALARY_RANGES: dict[str, tuple[int, int]] = {
+SALARY_RANGES: dict[str, tuple[float, float]] = {
     "junior":       (4000.0,   7000.0),
     "pleno":        (8000.0,  14000.0),
     "senior":       (14000.0, 22000.0),
